add 10 problem solution + refactoring type of result

+++ ProblemSolutions/Problem_1.fs
@@ -10,4 +10,4 @@ let public GetResult =
     let limit = 999;
     let three = 3;
     let five = 5;
-    sumDevidableByX(three, 1000) + sumDevidableByX(five, limit) - sumDevidableByX(three * five, limit)
+    (uint64) (sumDevidableByX(three, 1000) + sumDevidableByX(five, limit) - sumDevidableByX(three * five, limit))
